app/probe_relighting/requirements.txt
torch
torchvision
pyyaml
streamlit
tqdm
test-tube
pytorch-lightning
